Tests/LibWeb/Text/expected/form-has-invalid-style-invalidation.txt

with invalid child: rgb(0, 0, 255)
with valid child: rgb(255, 0, 0)
